notebooks/phase4_deployment/requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
requests
python-dotenv
pytest
httpx
